перенесла closePhoto и openModal
