+++ js/create-dom.js
@@ -1,21 +1,5 @@
 import {ArrayofPhoto} from './data.js';
-import{createBigPhoto,bigPhoto} from './big-photo.js';
-
-const closePhoto = document.querySelector('.big-picture__cancel');
-
-closePhoto.addEventListener('click',() => {
-  bigPhoto.classList.add('hidden');
-});
-document.addEventListener('keydown',(evt)=>{
-  if(evt.key === 'Escape'){
-    evt.preventDefault();
-    bigPhoto.classList.add('hidden');
-  }
-});
-const openModal = ()=>{
-  bigPhoto.classList.remove('hidden');
-  document.body.classList.add('modal-open');
-};
+import{createBigPhoto,openModal} from './big-photo.js';
 //список куда вставлять
 const picContainer = document.querySelector('.pictures');
 //доступ к шаблону списка фотографий
